Add bin and src structure

+++ src/train_mobilenet.py
@@ -4,20 +4,19 @@ import os
 import shutil
 from datetime import datetime
 
-import tensorflow as tf
 from sklearn.preprocessing import LabelBinarizer
 from tensorflow.keras.applications.mobilenet_v2 import MobileNetV2
-from tensorflow.keras.callbacks import ModelCheckpoint
-from tensorflow.keras.layers import Input
+from tensorflow.keras.callbacks import EarlyStopping, ModelCheckpoint
 from tensorflow.keras.models import Model
 from tensorflow.keras.optimizers import SGD, RMSprop
 
 from custom_dataset_loader_and_splitter import CustomDatasetLoaderAndSplitter
+from data_generator import DataGenerator
 from fcheadnet import FCHeadNet
 
 
 def build_output_folder(clear_output) -> str:
-    if clear_output:
+    if clear_output and os.path.exists("models"):
         shutil.rmtree("models")
 
     if not os.path.exists("models"):
@@ -37,7 +36,10 @@ def train_mobilenet(args, output_path):
     # load the images and get the splits
     class_names = list(os.listdir(args.dataset_path))
     data_loader = CustomDatasetLoaderAndSplitter(
-        args.dataset_path, validation=args.validation_split, test=args.test_split
+        args.dataset_path,
+        validation=args.validation_split,
+        test=args.test_split,
+        seed=args.seed,
     )
 
     if args.test_split > 0:
@@ -48,8 +50,16 @@ def train_mobilenet(args, output_path):
     # convert the labels from integers to vectors
     train_y = LabelBinarizer().fit_transform(train_y)
     val_y = LabelBinarizer().fit_transform(val_y)
+
+    # tensorflow tries to allocate the whole dataset to the GPU
+    # so we need to provide a DataGenerator to feed the data in batches
+    train_gen = DataGenerator(train_x, train_y, args.batch_size)
+    val_gen = DataGenerator(val_x, val_y, args.batch_size)
+
+    # set up test data if needed
     if args.test_split > 0:
         test_y = LabelBinarizer().fit_transform(test_y)
+        test_gen = DataGenerator(test_x, test_y, args.batch_size)
 
     # load the pre-trained network, ensuring the head FC layer sets are left off (include_top=False)
     print("[INFO] loading model...")
@@ -85,11 +95,9 @@ def train_mobilenet(args, output_path):
     # typical warmup are 10-30 epoch
     print("[INFO] training head...")
     history = model.fit(
-        train_x,
-        train_y,
-        validation_data=(val_x, val_y),
-        epochs=5,
-        batch_size=args.batch_size,
+        train_gen,
+        validation_data=val_gen,
+        epochs=20,
         verbose=1,
     )
 
@@ -116,17 +124,16 @@ def train_mobilenet(args, output_path):
     checkpoint = ModelCheckpoint(
         filepath, monitor="val_accuracy", verbose=1, save_best_only=True, mode="max"
     )
-    callbacks_list = [checkpoint]
+    early_stop_callback = EarlyStopping(monitor="val_loss", patience=5)
+    callbacks_list = [checkpoint, early_stop_callback]
 
     # train the model again, this time fine-tuning *both* the final set
     # of CONV layers along with our set of FC layers
     print("[INFO] fine-tuning model...")
     history = model.fit(
-        train_x,
-        train_y,
-        validation_data=(val_x, val_y),
-        epochs=5,
-        batch_size=args.batch_size,
+        train_gen,
+        validation_data=val_gen,
+        epochs=50,
         verbose=1,
         callbacks=callbacks_list,
     )
@@ -152,6 +159,7 @@ if __name__ == "__main__":
         help="Output name of the trained model.",
     )
     parser.add_argument(
+        "-b",
         "--batch-size",
         type=int,
         default=32,
@@ -174,7 +182,15 @@ if __name__ == "__main__":
         action="store_true",
         help="Remove all subfolders in output models/ folder. Warning, this is not reversible!",
     )
+    parser.add_argument(
+        "-s",
+        "--seed",
+        type=int,
+        default=-1,
+        help="Seed used to randomly split the dataset.",
+    )
     args = parser.parse_args()
 
     output_path = build_output_folder(args.clear_models)
+
     train_mobilenet(args, output_path)
